Close #206 - separate `MetadataFactory`
